persisting ready but disabled
