+++ app/components/DashboardCard.jsx
@@ -45,5 +45,3 @@ export default class DashboardCard extends Component{
   )
   }
 }
-//chartGenerator(userCode)
-
