+++ app/components/DashboardView.jsx
@@ -14,7 +14,6 @@ export default class DashboardView extends Component {
     render(){
         return (
             <div className = "container">
-            <p>Random fucking words</p>
                 <button onClick={this.print} className="pull-right" id="react-no-print"><span className="glyphicon glyphicon-download"></span></button>
 
                 <PageHeader header="Dashboard"/>
