+++ renderer.js
@@ -20,3 +20,4 @@ window.ReactFauxDOM = ReactFauxDOM
 
 let d3 = require('d3')
 window.d3 = d3
+
